Common: Move get_all_regular_files_in_directory() to utils.file_utils
